+++ Statemachine_ST7783_ADC/main.h
@@ -6,5 +6,5 @@
 
 #define CHARTONUM(a, b) ((a - '0') * b)
 
-void bin2asc(unsigned int value, char buffer[], byte digits);
+void bin2asc(uint16_t value, char buffer[], uint8_t digits);
